requirements.txt
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-generator==1.10
attrs==22.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
exceptiongroup==1.1.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
h11==0.14.0
idna==3.4
ipykernel @ file:///C:/ci/ipykernel_1646982785443/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651601994004/work
jedi @ file:///C:/ci/jedi_1644315428289/work
jupyter-client @ file:///C:/ci/jupyter_client_1650515989971/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656286743/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929390/work
numpy==1.24.1
outcome==1.2.0
packaging==23.0
pandas==1.5.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser==2.21
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.21.0
pytz==2022.7
pywin32==302
pyzmq @ file:///C:/ci/pyzmq_1638435148211/work
requests==2.28.1
selenium==4.7.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webdriver-manager==3.8.5
wincertstore==0.2
wsproto==1.2.0
